# query/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server Classroom service",
        "title": "Classroom API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/c/classroom",
    "paths": {
...
        "/works/delete_attached/{idWork}/{idAttached}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Delete Attached",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idWork",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idAttached",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad path param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No existe este elemento adjunto al trabajo",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/works/delete_file_work/{idWork}/{idFile}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works",
                    "classroom",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Delete file classroom",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idWork",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idFile",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "No se puede eliminar un trabajo calificado",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No se encontró el archivo a eliminar en este trabajo",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/works/delete_item_pattern/{idWork}/{idItem}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Delete Item Pattern",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idWork",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idItem",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad path param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No existe el item a eliminar",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/works/delete_work/{idWork}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Delete work",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idWork",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "No se puede eliminar un trabajo calificado",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/works/finish_form/{idWork}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "works",
                    "classroom",
                    "roles.student",
                    "roles.student_directive"
                ],
                "summary": "Finish form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idWork",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Desc",
                        "name": "answers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.AnswersForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "forms.Answer": {
            "type": "object",
            "required": [
                "question"
            ],
            "properties": {
                "answer": {
                    "type": "integer",
                    "example": 1
                },
                "question": {
                    "type": "string"
                },
                "response": {
                    "type": "string",
                    "example": "Response..."
                }
            }
        },
...
        "forms.AnswersForm": {
            "type": "object",
            "required": [
                "answers"
            ],
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.Answer"
                    }
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}